requirements.txt
bcrypt==3.1.7
cffi==1.14.1
configparser==5.0.0
pkg-resources==0.0.0
pycparser==2.20
PyMySQL==0.10.0
pyotp==2.4.0
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.15.0
PyQt5-sip==12.8.0
six==1.15.0
